refactor: extract proposal creation helpers
